test: 統合テストとCLIテストを分離
