+++ package.json
@@ -9,7 +9,7 @@
   "scripts": {
     "test": "yarn test:unit && yarn test:integration",
     "test:unit": "node --experimental-vm-modules node_modules/jest/bin/jest.js --testPathIgnorePatterns=test/",
-    "test:integration": "node --test test/cli.test.mjs",
+    "test:integration": "node --test test/cli.test.mjs test/integration.test.mjs",
     "test:generate-fixtures": "node test/generate-fixtures.mjs"
   },
   "author": "DenkiYagi Inc.",
